perf(comms): int16-native Opus hot path (Phase 5)
Phase 5 of the comms refactor (see .claude/plans/comms-refactor.md) moves
the TX and RX audio hot paths to int16 end-to-end, eliminating float32
conversions that hit softfloat on mipsle targets:

codec:
- AudioEncoder / AudioDecoder gain EncodeS16 / DecodeS16 int16-native
  entry points and Close() methods. The legacy Encode / Decode /
  DecodeFloat32 methods remain as deprecated shims for migration.
- Short-buffer guards return codec.ErrShortBuffer. PLC is routed through
  libopus DecodePLC (int16) rather than DecodePLCFloat32.

TX hot path (broadcast_encoder.go):
- PortAudio capture callback signature is now []int16, so PortAudio
  delivers samples in paInt16 format (the gordonklaus binding selects
  the C sample format from the callback signature via reflection).
- encCh carries *[]int16 pooled frames from int16Pool; the float32
  intermediate buffer and float32-to-int16 conversion loop are gone.
- Mic gain is applied in-place in int16 space with int16 clipping.
- encodeOne calls EncodeS16 directly. Pool release via defer on every
  path including encode-error returns.

RX hot path (receive.go / comms.go):
- playoutOneFrame takes []int16 and calls DecodeS16 for both real
  frames and PLC. The PortAudio playback callback signature is []int16,
  so playback opens with paInt16 directly.
- Beep start/stop tones are generated as []int16 (amplitude 0.2 * 32767)
  and flow through the int16 playbackBuffer unchanged.

Boundary conversion:
- ROIP VOX tap is the one remaining float32 consumer. The broadcast
  encoder converts int16 -> float32 at the tap boundary only when a tap
  is registered (ROIP + VOX only). Regular TX (openvlm / nanoptt / web)
  never enters this branch. The ROIP VOX threshold semantics are
  unchanged because the conversion factor (/32768) matches the original
  PortAudio float32 normalization.
- The WebAudioBridge is Opus-bytes-only and not affected.

Tests updated in lockstep: mockEncoder / mockDecoder gain EncodeS16 /
DecodeS16 / Close; broadcast_encoder_test, receive_test, transmit_test,
multiport_test, comms_test, bench_test, integration_test all use []int16.
New codec tests cover EncodeS16 / DecodeS16 round-trip, PLC, short-buffer
errors, and Close idempotency.

Hot-path benchmarks (amd64 / Xeon @ 2.1 GHz, 1s):
  BenchmarkDecodeOpusS16          29201 ns/op   0 B/op   0 allocs/op
  BenchmarkMicGainInt16            1871 ns/op   0 B/op   0 allocs/op
  BenchmarkPlayoutOneFrame_Mock     167 ns/op   0 B/op   0 allocs/op
  BenchmarkPlayoutOneFrame_Real     172 ns/op   0 B/op   0 allocs/op
  BenchmarkPlayoutOneFrame_PLC       88 ns/op   0 B/op   0 allocs/op

Zero allocations on every hot-path benchmark. Pre-Phase-5 baseline is
not captured in this commit because the legacy float32 path is gone;
the architectural win on mipsle is elimination of softfloat conversions
(~frameSize multiplies + bound checks per frame, per direction).

Verification:
- go test ./internal/comms/...              PASS
- go test -race ./internal/comms/...        PASS
- go vet ./internal/comms/...               PASS

RTP wire format and audible output are unchanged; the swap is internal
to the codec and audio subsystems.

+++ internal/comms/receive_test.go
@@ -18,7 +18,7 @@ func newReceiveRuntime() (*CommsRuntime, *portChannel) {
 	}
 	pc.sendEnabled.Store(true)
 	pc.receiveEnabled.Store(true)
-	pc.playbackBuffer = make(chan []float32, 4)
+	pc.playbackBuffer = make(chan []int16, 4)
 	rt := &CommsRuntime{
 		ports:   []*portChannel{pc},
 		decoder: &mockDecoder{returnN: int(rtpFrameSamples)},
@@ -29,7 +29,7 @@ func newReceiveRuntime() (*CommsRuntime, *portChannel) {
 
 // isAllZero reports whether every sample in the slice is exactly zero.
 // Used by playoutOneFrame tests to distinguish silence from decoded audio.
-func isAllZero(out []float32) bool {
+func isAllZero(out []int16) bool {
 	for _, v := range out {
 		if v != 0 {
 			return false
@@ -43,8 +43,8 @@ func isAllZero(out []float32) bool {
 
 // driveOneFrame is a small helper for tests that need to call playoutOneFrame
 // against a fresh PCM output buffer of the standard frame size.
-func driveOneFrame(cfg *CommsConfig, pc *portChannel, rt *CommsRuntime, jb *rtpJitterBuffer) []float32 {
-	out := make([]float32, frameSize)
+func driveOneFrame(cfg *CommsConfig, pc *portChannel, rt *CommsRuntime, jb *rtpJitterBuffer) []int16 {
+	out := make([]int16, frameSize)
 	cfg.playoutOneFrame(pc, rt, jb, out)
 
 	return out
@@ -281,7 +281,7 @@ func TestReceiveLoop_DropsOwnPackets(t *testing.T) {
 	}
 	pc.sendEnabled.Store(true)
 	pc.receiveEnabled.Store(true)
-	pc.playbackBuffer = make(chan []float32, 16)
+	pc.playbackBuffer = make(chan []int16, 16)
 	rt := &CommsRuntime{
 		ports:   []*portChannel{pc},
 		decoder: &mockDecoder{returnN: int(rtpFrameSamples)},
@@ -337,7 +337,7 @@ func TestReceiveLoop_DropsMalformedRTP(t *testing.T) {
 	}
 	pc.sendEnabled.Store(true)
 	pc.receiveEnabled.Store(true)
-	pc.playbackBuffer = make(chan []float32, 8)
+	pc.playbackBuffer = make(chan []int16, 8)
 	rt := &CommsRuntime{
 		ports:   []*portChannel{pc},
 		decoder: &mockDecoder{},
@@ -422,7 +422,7 @@ func TestReceiveLoop_StampsLastRemoteRx(t *testing.T) {
 	}
 	pc.sendEnabled.Store(true)
 	pc.receiveEnabled.Store(true)
-	pc.playbackBuffer = make(chan []float32, 32)
+	pc.playbackBuffer = make(chan []int16, 32)
 	rt := &CommsRuntime{
 		ports:   []*portChannel{pc},
 		decoder: &mockDecoder{returnN: int(rtpFrameSamples)},
@@ -510,7 +510,7 @@ func TestReceiveLoop_SocketSwapResetsJitter(t *testing.T) {
 	}
 	pc.sendEnabled.Store(true)
 	pc.receiveEnabled.Store(true)
-	pc.playbackBuffer = make(chan []float32, 64)
+	pc.playbackBuffer = make(chan []int16, 64)
 
 	rt := &CommsRuntime{
 		ports:   []*portChannel{pc},
